Issues when MAXWRAPLENGTHCONST, allow to set using make
Makefile:
- Renamed LD_FLAGS to LDFLAGS for proper Makefile conformance.
- More work still needs to be done to make this a useful Makefile.

trf30dat.h:
- MAXWRAPLENGTHCONST might be too high on some machines, and at very
large values, won't work with the 32-bit executable at all.
- Use an ifndef guard so that we can define it when building with
make (or other system) and have a fallback when it is not defined.

trfrun.h:
- Just made one change in how a variable is dynamically allocated and
zeroed. We control the compilation, so we're free to use the latest
C standard compiler. This means we don't need to cast the return
values of malloc/calloc/etc. And we can just use calloc instead of
malloc+memset. If this causes issues compiling on a certain platform,
we'll handle it. I might change other allocations if it fixes issues
on certain platforms.