Removed compile flag for MAXWRAPLENGTH, now an option, using getopt
Makefile:
- Removed different values for the MAXWRAPLENGTHCONST macro that was
being defined for each platform. This is replaced by a user-set
variable (see changelog for trf.c below)
- Since the above no longer varies by word length, Apple-specific
CFLAGS have been merged to make a universal binary (NEEDS TESTING)
- Similarly, no more word-length-dependent variables for win/unix

trf.c:
- Now using getopt for ease of new option addition and for options
with agruments
- Updated documentation for new option

trf30dat.c, trf30dat.h:
- Changed all instances of MAXWRAPLENGTH with paramset.maxwraplength
- Bandcenter is now dynamically allocated because this array depended
on the macro MAXWRAPLENGTHCONST being defined. It could not be
inizialized using MAXWRAPLENGTH or paramset.maxwraplength since these
are undefined during compile time. Now, Bandcenter is still a global
variable, but in trf30dat.c we check if it is defined, and if not,
it is allocated. It is never freed since the allocation only happens
once. This should be fine since the array was used as a shared value
before being changed to being dynamically allocated.

trf30dat.h:
- Added member, maxwraplength, to paramset struct for new option.

trfrun.h:
- Added an fclose to quiet cppcheck complaint of a resource leak from
not closing a file handle on an error.
- Added a check for word size when printing memory allocation error
message, since the return type of sizeof varies by platform word
length.