Changed alloc/free of Bandcenter to help fix SIGKILL
tr30dat.c:
- Bandcenter is now allocated at the start of the newtupbo function and
freed at the end of it. Since we know what the value of maxwraplength
should be at this point, and since Bandcenter is shared between at
least two functions called in newtupbo, this *should* be OK. Testing
to see if memory is properly allocated and freed to avoid triggering
the Linux kernel's oom-killer.